+++ index.js
@@ -1,4 +1,4 @@
-// version: v1.0.27 | build: 2026-07-25T20:47:22Z
+// version: v1.0.28 | build: 2026-07-25T20:55:20Z
 let products = [], orders = [], customers = [];
 let settings = {};
 let cQty = {};